% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{auth_key}
\alias{auth_key}
\title{store API key in global environment}
\usage{
auth_key(api_key='')
}
\arguments{
\item{api_key}{A unque key given by TMDb for each account}
}
\value{
An character object stored in the global environment
}
\description{
This is a function that store API key for the sake of functions 
    in the pakage
}

