% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discover_query_strings}
\alias{discover_query_strings}
\title{Legit query element for the discover functionality}
\format{A list of two elements
\describe{
  \item{movie}{a group of legit elements to include in query list if search_cat
  is movie}
  \item{tv}{a group of legit elements to include in query list if search_cat
  is tv}
}}
\source{
\url{https://developers.themoviedb.org/3/discover/movie-discover}
}
\usage{
discover_query_strings
}
\description{
Legit query element for the discover functionality
}
\keyword{datasets}

