% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_functions.R
\name{get_latest_person}
\alias{get_latest_person}
\title{Get the most newly created person}
\usage{
get_latest_person(language='en-US')
}
\arguments{
\item{language}{Language of movie. By default is 'en-US'}
}
\value{
An object of class \code{tmdb_api} that consists quried data in a 
    data frame or a list and a metadata list
}
\description{
This function returns the most newly created person. This is a live 
    response and will continuously change.
}
\examples{
\dontrun{
result_data <- get_latest_person(language='en-US')
}
}

