% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movie_functions.R
\name{get_now_playing_movie}
\alias{get_now_playing_movie}
\title{Get the "now playing" movies}
\usage{
get_now_playing_movie(language='en-US', page=1)
}
\arguments{
\item{language}{Language of movie. By default is 'en-US'}

\item{page}{Which page to return}
}
\value{
An object of class \code{tmdb_api} that consists quried data in a 
    data frame or a list and a metadata list
}
\description{
This function fetch the "now playing" movies. This is a date based 
    query looking at the primary_release_date of a movie.
}
\examples{
\dontrun{
result_data <- get_now_playing_movie(language='en-US', page=2)
}
}

