% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_functions.R
\name{get_person_detail}
\alias{get_person_detail}
\title{Get detailed information of a certain person}
\usage{
get_person_detail(person_id, person_name=NULL, detail_cat=NULL, 
                        language='en-US', append_to_response=NULL)
}
\arguments{
\item{person_id}{The person_id of the person of interest}

\item{person_name}{If no person_id is provided, use person_name to match and 
identify the person}

\item{detail_cat}{The variable that clarify the query category. For full list
of legit values, print person_detail_cat_list}

\item{language}{language of movie. By default is 'en-US'}

\item{append_to_response}{A series of strings seperated by comma that clearify
objects to append that are related to this quried person. Possible values 
are imgages and videos}
}
\value{
An object of class \code{tmdb_api} that consists quried data in a 
    data frame or a list and a metadata list
}
\description{
This function takes in a person_id and a detail category to return.
    Noted that this function allow query by person_id or by person name. If by 
    name, two queries need to be sent to API, the first to get person_id, the 
    second to request data by person_id
}
\examples{
\dontrun{
result_data <- get_person_detail(person_id=287, detail_cat='movie_credits')
}
}

