% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_functions.R
\name{get_popular_person}
\alias{get_popular_person}
\title{Get the list of popular people on TMDb}
\usage{
get_popular_person(language='en-US', page=1)
}
\arguments{
\item{language}{Language of movie. By default is 'en-US'}

\item{page}{Which page to return}
}
\value{
An object of class \code{tmdb_api} that consists quried data in a 
    data frame or a list and a metadata list
}
\description{
This function fetch the list of popular people on TMDb. This 
    list updates daily.
}
\examples{
\dontrun{
result_data <- get_popular_person(language='en-US', page=2)
}
}

