% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{get_result}
\alias{get_result}
\title{Query data using a URL}
\usage{
get_result(url)
}
\arguments{
\item{url}{A TMDb url string to pull data from. url can be get using function 
in this package such as get_discover_url and get_search_url}
}
\value{
An object of class \code{tmdb_api}. The object consists of two elements. 
   The first element called data_frame is a \code{data frame} if the requested 
   returnings can be easily converted to data frame witout confusion. The second 
   element called response_detail is a \code{list} that contains metadata of the
   query such as headers, url and status code. Raw data in json format is 
   also available in response_detail.
}
\description{
This is a function that takes an assembled URL and query data from it, especially for the query where users specify which page to return. This function will generate message of total page so the user can modify their query accordingly.
}
\examples{
\dontrun{
url <- 'https://api.themoviedb.org/3/discover/movie?api_key={Put_API_Key_Here}&year=2011&with_genres=12,18&language=en-US&page=14'
result <- get_result(url)
}

}
\keyword{API}
\keyword{TMDb}

