% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{get_search_url}
\alias{get_search_url}
\title{Get url to pull data from using the search functionality}
\usage{
get_search_url(query_list=list(), search_cat='movie')
}
\arguments{
\item{query_list}{The list of criterion that narrow the range of returned results. 
To see the full list of legit search categories please print 
search_query_string}

\item{search_cat}{The variable that clarify search category}
}
\value{
An url in format of character.
}
\description{
This function generate url mainly based on one or a series 
    of keywords users specify to query from one of various target categories. 
    To see the full list of legit search categories please print 
    search_query_string to see
}
\examples{
\dontrun{
sample_query_list <- list(language='en-US', query='kill bill')
target_url <- get_search_url(sample_query_list, 'movie')
}
}

