% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movie_detail_cat_list}
\alias{movie_detail_cat_list}
\title{Legit detail_cat values for the function get_movie_detail()}
\format{A string vector
\describe{
    A series of values that are legit for the detail_cat argument in the 
    function get_movie_detail()
}}
\source{
\url{https://developers.themoviedb.org/3/movies/get-movie-details}
}
\usage{
movie_detail_cat_list
}
\description{
Legit detail_cat values for the function get_movie_detail()
}
\keyword{datasets}

