% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{person_detail_cat_list}
\alias{person_detail_cat_list}
\title{Legit query element for the function get_person_detail()}
\format{A string vector
\describe{
    A series of values that are legit for the detail_cat argument in the 
    function get_person_detail()
}}
\source{
\url{https://developers.themoviedb.org/3/movies/get-movie-details}
}
\usage{
person_detail_cat_list
}
\description{
Legit query element for the function get_person_detail()
}
\keyword{datasets}

