% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{print.tmdb_api}
\alias{print.tmdb_api}
\title{Print object tmdb_api}
\usage{
print.tmdb_api(rst) or rst
}
\arguments{
\item{rst}{The returned object of function get_result() or get_result_general().
rst should be of class \code{tmdb_api}}
}
\value{
An object of class \code{tmdb_api}.
}
\description{
This function controls the output when users print the response object
}
\examples{
\dontrun{
url <- 'https://api.themoviedb.org/3/discover/movie?api_key={Put_API_Key_Here}&year=2011&with_genres=12,18&language=en-US'
result <- get_result_general(url)
result
}
}

