% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{search_query_string}
\alias{search_query_string}
\title{Legit query element for the function get_search_url() and search_multi_page()}
\format{A list of seven elements
\describe{
  \item{company}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is company}
  \item{collection}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is collection}
  \item{keyword}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is keyword}
  \item{movie}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is movie}
  \item{multi}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is multi}
  \item{person}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is person}
  \item{tv}{a group of legit elements to include in query list of function 
  get_search_url() and search_multi_page() if search_cat is tv}
}}
\source{
\url{https://developers.themoviedb.org/3/discover/movie-discover}
}
\usage{
search_query_string
}
\description{
Legit query element for the function get_search_url() and search_multi_page()
}
\keyword{datasets}

