% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\docType{data}
\name{TMDb_Env}
\alias{TMDb_Env}
\title{Create an environment to hold api_key}
\format{An object of class \code{environment} of length 0.}
\usage{
TMDb_Env
}
\description{
Create an environment to hold api_key
}
\keyword{datasets}

